requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
streamlit==1.42.0
pandas==2.2.3
scipy==1.15.1
statsmodels==0.14.4
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
